+++ backend/src/main/java/com/prinpedia/backend/repository/TagRepository.java
@@ -1,0 +1,8 @@
+package com.prinpedia.backend.repository;
+
+import com.prinpedia.backend.entity.Tag;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface TagRepository extends MongoRepository<Tag, Integer> {
+    Tag findByKeyword(String keyword);
+}
