+++ backend/src/main/java/com/prinpedia/backend/service/TagService.java
@@ -1,0 +1,8 @@
+package com.prinpedia.backend.service;
+
+import java.util.List;
+
+public interface TagService {
+    List<String> processSearch(String searchText);
+    List<String> findTitles(List<String> keywordList);
+}
