--- src/main/java/com/prinpedia/backend/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.prinpedia.backend.repository;
-
-import com.prinpedia.backend.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-    User findByUsername(String username);
-}
-
