--- src/main/java/com/prinpedia/backend/service/TagService.java
@@ -1,8 +1,0 @@
-package com.prinpedia.backend.service;
-
-import java.util.List;
-
-public interface TagService {
-    List<String> processSearch(String searchText);
-    List<String> findTitles(List<String> keywordList);
-}
